Update code to use size_t
The`size_t` type is used to represent the size of an object.
It is defined as the return type of the sizeof operator. As an
implication size_t is used for holding an array index.

Many of the functions that would naturally return `size_t`
also return -1 when an error has occured. To hold the error
these functions still return int or long.

Being consistant with the usage of `size_t` helps reduce the
amount of casting when compiling for systems that are not
32-bit.

This is a first pass through the code to identify locations
that `int` or `long` was used when `size_t` probably should
have been used.

This is not an exastive update.

+++ api/oc_rep.c
@@ -55,7 +55,7 @@ oc_rep_finalize(void)
   }
   if (g_err != CborNoError)
     return -1;
-  return (int)size;
+  return size;
 }
 
 static oc_rep_t *
